Display lists of authors properly on: homepage, tag pages, and article pages
