+++ lib/articles.js
@@ -17,7 +17,9 @@ const LIST_ARTICLES = `
         headline
         byline
         authors {
+          id
           name
+          slug
         }
         content
         firstPublishedOn
@@ -40,7 +42,9 @@ const LIST_ARTICLES_BY_SLUG = `
         byline
         slug
         authors {
+          id
           name
+          slug
         }
         content
         tags {
@@ -315,6 +319,11 @@ const GET_ARTICLE = `
         tags {
           title
         }
+        authors {
+          id
+          name
+          slug
+        }
         firstPublishedOn
         lastPublishedOn
         searchTitle
@@ -335,6 +344,7 @@ const GET_ARTICLE_BY_SLUG = `
         id
         slug
         authors {
+          id
           name
           slug
         }
